--- src/main/java/no/gunbang/market/common/exception/ForbiddenException.java
@@ -1,9 +1,0 @@
-package no.gunbang.market.common.exception;
-
-public class ForbiddenException extends BaseException {
-
-    public ForbiddenException(ErrorCode errorCode) {
-        super(errorCode.getMessage(), errorCode.getStatus());
-    }
-
-}
