--- src/main/java/no/gunbang/market/common/exception/UnauthorizedException.java
@@ -1,9 +1,0 @@
-package no.gunbang.market.common.exception;
-
-public class UnauthorizedException extends BaseException {
-
-    public UnauthorizedException(ErrorCode errorCode) {
-        super(errorCode.getMessage(), errorCode.getStatus());
-    }
-
-}
