+++ src/main/java/no/gunbang/market/domain/market/dto/response/MarketPopularResponseDto.java
@@ -1,4 +1,4 @@
-package no.gunbang.market.domain.market.dto;
+package no.gunbang.market.domain.market.dto.response;
 
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
