+++ src/main/java/no/gunbang/market/domain/user/service/UserHistoryService.java
@@ -5,8 +5,8 @@ import lombok.RequiredArgsConstructor;
 import no.gunbang.market.domain.auction.dto.response.AuctionHistoryResponseDto;
 import no.gunbang.market.domain.auction.dto.response.BidHistoryResponseDto;
 import no.gunbang.market.domain.auction.repository.AuctionRepository;
-import no.gunbang.market.domain.market.dto.MarketHistoryResponseDto;
-import no.gunbang.market.domain.market.dto.TradeHistoryResponseDto;
+import no.gunbang.market.domain.market.dto.response.MarketHistoryResponseDto;
+import no.gunbang.market.domain.market.dto.response.TradeHistoryResponseDto;
 import no.gunbang.market.domain.market.repository.MarketRepository;
 import org.springframework.stereotype.Service;
