+++ src/test/java/no/gunbang/market/TradeMarketTest.java
@@ -15,8 +15,8 @@ import no.gunbang.market.common.entity.InventoryRepository;
 import no.gunbang.market.common.entity.Item;
 import no.gunbang.market.common.entity.ItemRepository;
 import no.gunbang.market.common.entity.Status;
-import no.gunbang.market.domain.market.dto.MarketTradeRequestDto;
-import no.gunbang.market.domain.market.dto.MarketTradeResponseDto;
+import no.gunbang.market.domain.market.dto.request.MarketTradeRequestDto;
+import no.gunbang.market.domain.market.dto.response.MarketTradeResponseDto;
 import no.gunbang.market.domain.market.entity.Market;
 import no.gunbang.market.domain.market.entity.Trade;
 import no.gunbang.market.domain.market.repository.MarketRepository;
